Create sortingMarbles.js
Grouping marbles is a number game iS that involves arranging different sets of marbles by their size. More formally, a player is given n marbles, each having : unique number from 1 to n associated with it. Initially, these marbles are divided into 3 groups groupA, groupB, and groupC of sizes n1, n2, and n3 respectively. Arrange them in such a way that if we sort each individual group independently in increasing order and create a list by appending to the list:
- All the elements of groupA from smallest to largest
- Followed by all the elements of groupB from smallest to largest
- Followed by all elements of groupC from smallest to largest then, the resultant list contains all numbers from 1 to n sorted in increasing order.

This can be done through a series of moves. In one move, the player moves one marble from one group to some other group. The task is to return the minimum number of moves in which the player can obtain the desired arrangement. Note: The sizes of the 3 groups need not remain the same after the series of moves. Also, it is possible that after the series of moves, some groups do not have any marbles. It is guaranteed that initially, each marble is part of one and only one group.